refactor: 聚合excel工具类
